GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Dev (refs/gitbutler/Dev)
   branch head: 38944ce68524efbcd2e68b947d5c8b8f508d7e9b
   - package.json
 - Virtual branch (refs/gitbutler/Virtual-branch)
   branch head: c87fdd2f57a9dd6655be9f5d73abe982c5bc2b0d
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch